[ENG-495] - Reorganized for dotnet core (netstandard)
This removes some of the portability directives
in favour of the new dotnet tooling.

It also drops use of normaliz.dll in favour of
String.Normalize.

There were no tests included with the original
code...so, none here either. Would have been
really nice to see some to compare the output
but as I don't have any historical reasons for
ensuring the implementation is identical I have
been lazy. This code works when imported.